chore : move dummyData.ts
